+++ portal/src/main/webapp/js/src/oncoprint/Oncoprint.js
@@ -257,6 +257,9 @@ define("Oncoprint",
                     var data = clinicalData.concat(params.geneData);
                 }
 
+                
+                var Sampledata = params.geneData;
+
                 var clinical_attrs = params.clinical_attrs      // extract attr_ids
                     .map(function(attr) { return attr.attr_id; });
 
@@ -265,7 +268,6 @@ define("Oncoprint",
                 clinical_attrs = clinical_attrs.filter(function(i) { return i !== undefined; });
                 clinical_attrs = clinical_attrs || params.clinical_attrs;
 
-//                var attributes = clinical_attrs.concat(params.genes);
                 var attributes = tracks.concat(params.genes);
 
                 data = utils.process_data(data, attributes);
@@ -344,6 +346,9 @@ define("Oncoprint",
                     } 
                 });
                 
+
+                Sampledata = utils.process_data(Sampledata, params.genes);
+
                 // keeps track of the order specified by the user (translates to vertical
                 // order in the visualization)
                 var attr2index = (function() {
@@ -1365,7 +1370,7 @@ define("Oncoprint",
                         sortBy: sortBy,
                         getPdfInput: getPdfInput,
                         getOncoprintData: function() {
-                            return data;
+                            return Sampledata;
                             //return newgenepaneldata
                         },
                         getOncoprintExcludedGenepanelData: function(){