+++ portal/src/main/webapp/js/src/oncoprint/main-boilerplate.js
@@ -46,7 +46,7 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
     var showalteredcaseflag = true;
     var showwhitespaceflag = true;
     var increasesortflag = true;
-    var showPatientIdflat = false;
+    var showPatients = true;
     // add in controls from template
     document.getElementById('oncoprint_controls').innerHTML
         = _.template(document.getElementById('main-controls-template').innerHTML)();
@@ -138,7 +138,7 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
     var extraAttributes=[]; // used to record attributes names add customized
     var sortStatus=[];
     var genepanelValues; // data structure store all data related to genepanel values include genepanel attribute data and genepanel data
-    
+
     var cases = window.PortalGlobals.getCases();
     var genes = window.PortalGlobals.getGeneListString().split(" ");
 
@@ -179,10 +179,19 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
         success: function(data) {
             //for add clinical attributes in url to data start
             var urlValueNow = window.location.href;//get current page url
-            var beginOfDeleteElementNow = urlValueNow.indexOf('&clinicallist=');
+            var beginOfDeleteElementNow = urlValueNow.indexOf('clinicallist=');
+            
+            if(urlValueNow.indexOf("&show_samples=true")>-1)
+            {
+                showPatients =false;
+            }
+            
             if(beginOfDeleteElementNow > 0)
             {
-                var clinicallistStringNow = urlValueNow.substring(beginOfDeleteElementNow + ('&clinicallist=').length,urlValueNow.length);
+//                var clinicallistStringNow = urlValueNow.substring(beginOfDeleteElementNow + ('&clinicallist=').length,urlValueNow.length);
+                var clinicallistStringNow = urlValueNow.substring(beginOfDeleteElementNow);
+                clinicallistStringNow = (clinicallistStringNow.split("&"))[0];
+                clinicallistStringNow = (clinicallistStringNow.split("="))[1];
                 var clinicallistArrayNow = clinicallistStringNow.split('+');
 
                 var clinicalAttributeArray = [];
@@ -374,7 +383,8 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                             legend: document.getElementById('oncoprint_legend'),
                             sortStatus:sortStatus,
                             mutationColor:mutationColorControl
-                            },extraTracks,genepanelValues);
+                            },extraTracks,genepanelValues,showPatients);
+
 
                             outer_loader_img.hide();
                             $('#oncoprint #everything').show();
@@ -397,6 +407,22 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
 
                             zoom = reset_zoom();
                             invokeDataManager(); 
+
+                            if(showPatients)
+                            {
+                                var AlteredPatientsNum= calculatePatientNum(PortalGlobals.getAlteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                                var UnalteredPatientsNum= calculatePatientNum(PortalGlobals.getUnalteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                                var totalPatientsNum = _.union(AlteredPatientsNum,UnalteredPatientsNum);
+                                var percentOfAlteredPatients = Math.ceil((AlteredPatientsNum.length/totalPatientsNum.length * 100).toFixed(1));
+                                $('#altered_value').text("Altered in "+ AlteredPatientsNum.length + " ("+ percentOfAlteredPatients +"%) of "+ totalPatientsNum.length +" patients/cases");
+                            }
+                            else
+                            {
+                                $('#oncoprint_diagram_topatientid_icon img')[0].attributes.src.value = 'images/cool.svg';
+                                $('#switchPatientSample').text("Show patients in OncoPrint"); 
+                                $('#switchPatientSample')[0].attributes.valuetype.value = "samples";
+                                $('#altered_value').text("Altered in "+ PortalGlobals.getNumOfAlteredCases() + " ("+ Math.ceil(PortalGlobals.getPercentageOfAlteredCases()) +"%) of "+ PortalGlobals.getNumOfTotalCases() + " samples"); 
+                            }
                             
                         if($('#oncoprint_sortbyfirst_dropdonw span')[0].innerHTML === 'Sort by')
                         {
@@ -566,6 +592,23 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                         zoom = reset_zoom();
                         invokeDataManager(); 
                 });
+                zoom = reset_zoom();
+                invokeDataManager();                   
+                if(showPatients)
+                {
+                    var AlteredPatientsNum= calculatePatientNum(PortalGlobals.getAlteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                    var UnalteredPatientsNum= calculatePatientNum(PortalGlobals.getUnalteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                    var totalPatientsNum = _.union(AlteredPatientsNum,UnalteredPatientsNum);
+                    var percentOfAlteredPatients = Math.ceil((AlteredPatientsNum.length/totalPatientsNum.length * 100).toFixed(1));
+                    $('#altered_value').text("Altered in "+ AlteredPatientsNum.length + " ("+ percentOfAlteredPatients +"%) of "+ totalPatientsNum.length +" patients/cases");
+                }
+                else
+                {
+                    $('#oncoprint_diagram_topatientid_icon img')[0].attributes.src.value = 'images/cool.svg';
+                    $('#switchPatientSample').text("Show patients in OncoPrint"); 
+                    $('#switchPatientSample')[0].attributes.valuetype.value = "samples";
+                    $('#altered_value').text("Altered in "+ PortalGlobals.getNumOfAlteredCases() + " ("+ Math.ceil(PortalGlobals.getPercentageOfAlteredCases()) +"%) of "+ PortalGlobals.getNumOfTotalCases() + " samples"); 
+                }                
             }
         }
     });
@@ -617,10 +660,12 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                 positionOfDeleteElement = urlValue.indexOf('mutation'); //for delete it in url
             }
             
-            var beginOfDeleteElement = urlValue.indexOf('&clinicallist='); //for delete it in url
+            var beginOfDeleteElement = urlValue.indexOf('clinicallist='); //for delete it in url
             
-            var clinicallistString = urlValue.substring(beginOfDeleteElement+('&clinicallist=').length,urlValue.length);
-            var clinicallistArray = clinicallistString.split('+');
+//            var clinicallistString = urlValue.substring(beginOfDeleteElement+('&clinicallist=').length,urlValue.length);
+            var clinicallistString = urlValue.substring(beginOfDeleteElement);
+            clinicallistString = (clinicallistString.split("&"))[0];
+            var clinicallistArray = (clinicallistString.split("="))[1].split('+');
             if(clinicallistArray.length > 1)
             {
                 var DeleteElementAttr_id = recordDeleteElement.attr_id;
@@ -642,7 +687,8 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
             }
             else
             {
-                var newUrl = urlValue.slice(0,beginOfDeleteElement);
+//                var newUrl = urlValue.slice(0,beginOfDeleteElement);
+                var newUrl = urlValue.replace("&"+clinicallistString,'');
                 window.history.pushState({"html":window.location.html,"pageTitle":window.location.pageTitle},"", newUrl);
             }
             
@@ -770,7 +816,6 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                 mutationColor:mutationColorControl
             },extraTracks,genepanelValues,showPatientIdflat);
 
-//            oncoprint.sortBy(sortBy.val(), cases.split(" "));
             if($('#oncoprint_sortbyfirst_dropdonw span')[0].innerHTML === 'Sort by')
             {
                 oncoprint.sortBy("genes", cases.split(" "),mutationColorControl,mutationColorSort,sortStatus);
@@ -824,6 +869,7 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
             sortStatus:sortStatus,
             mutationColor:mutationColorControl
         },extraTracks,genepanelValues,showPatientIdflat);
+
         if(extraAttributes.length < 1)
         {
             $('#oncoprint-diagram-toolbar-buttons #clinical_first')[0].style.display = "none";
@@ -913,7 +959,7 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
         inner_loader_img.show();
         toggleControls(false); //disable toggleControls
 
-        inner_loader_img.hide();
+        inner_loader_img.hide()
         
         var topatientValue;
         if(topatient !== undefined)
@@ -1050,7 +1096,10 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                         }
                         else
                         {
-                            var newUrl = urlValue + "+mutation";
+                            var stringAfterClinical = urlValue.substring(positionOfSubmit);
+                            var stringArrayAfterClinical = stringAfterClinical.split("&");
+                            var clinicalString = stringArrayAfterClinical[0] + "+mutation";
+                            var newUrl = urlValue.replace(stringArrayAfterClinical[0],clinicalString);
                             window.history.pushState({"html":window.location.html,"pageTitle":window.location.pageTitle},"", newUrl);
                         }
                         
@@ -1070,6 +1119,18 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                         $('#oncoprint-diagram-removeWhitespace-icon img')[0].attributes.src.value = 'images/removeWhitespace.svg';
                         utils.make_mouseover(d3.selectAll('.sample rect'),{linkage:true});        // hack =(
                         invokeDataManager();
+                        if(showPatients)
+                        {
+                            var AlteredPatientsNum= calculatePatientNum(PortalGlobals.getAlteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                            var UnalteredPatientsNum= calculatePatientNum(PortalGlobals.getUnalteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                            var totalPatientsNum = _.union(AlteredPatientsNum,UnalteredPatientsNum);
+                            var percentOfAlteredPatients = Math.ceil((AlteredPatientsNum.length/totalPatientsNum.length * 100).toFixed(1));
+                            $('#altered_value').text("Altered in "+ AlteredPatientsNum.length + " ("+ percentOfAlteredPatients +"%) of "+ totalPatientsNum.length +" patients/cases");
+                        }
+                        else
+                        {
+                            $('#altered_value').text("Altered in "+ PortalGlobals.getNumOfAlteredCases() + " ("+ Math.ceil(PortalGlobals.getPercentageOfAlteredCases()) +"%) of "+ PortalGlobals.getNumOfTotalCases() + " samples"); 
+                        }    
                     }
                 });
             }
@@ -1102,7 +1163,6 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                             sortStatus:sortStatus,
                             mutationColor:mutationColorControl
                         },extraTracks,genepanelValues,showPatientIdflat);
-                             
 //                        oncoprint.sortBy(sortBy.val(), cases.split(" "));
                         if($('#oncoprint_sortbyfirst_dropdonw span')[0].innerHTML === 'Sort by')
                         {
@@ -1148,7 +1208,11 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                         }
                         else
                         {
-                            var newUrl = urlValue + "+FRACTION_GENOME_ALTERED";
+                            var stringAfterClinical = urlValue.substring(positionOfSubmit);
+                            var stringArrayAfterClinical = stringAfterClinical.split("&");
+                            var clinicalString = stringArrayAfterClinical[0] + "+FRACTION_GENOME_ALTERED";
+                            var newUrl = urlValue.replace(stringArrayAfterClinical[0],clinicalString);
+//                            var newUrl = urlValue + "+FRACTION_GENOME_ALTERED";
                             window.history.pushState({"html":window.location.html,"pageTitle":window.location.pageTitle},"", newUrl);
                         }
                         
@@ -1168,6 +1232,18 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                         $('#oncoprint-diagram-removeWhitespace-icon img')[0].attributes.src.value = 'images/removeWhitespace.svg';
                         utils.make_mouseover(d3.selectAll('.sample rect'),{linkage:true});        // hack =(
                         invokeDataManager();
+                        if(showPatients)
+                        {
+                            var AlteredPatientsNum= calculatePatientNum(PortalGlobals.getAlteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                            var UnalteredPatientsNum= calculatePatientNum(PortalGlobals.getUnalteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                            var totalPatientsNum = _.union(AlteredPatientsNum,UnalteredPatientsNum);
+                            var percentOfAlteredPatients = Math.ceil((AlteredPatientsNum.length/totalPatientsNum.length * 100).toFixed(1));
+                            $('#altered_value').text("Altered in "+ AlteredPatientsNum.length + " ("+ percentOfAlteredPatients +"%) of "+ totalPatientsNum.length +" patients/cases");
+                        }
+                        else
+                        {
+                            $('#altered_value').text("Altered in "+ PortalGlobals.getNumOfAlteredCases() + " ("+ Math.ceil(PortalGlobals.getPercentageOfAlteredCases()) +"%) of "+ PortalGlobals.getNumOfTotalCases() + " samples"); 
+                        }
                     }
                 });
             }
@@ -1246,7 +1322,11 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                         }
                         else
                         {
-                            var newUrl = urlValue + "+" + extraTracks[extraTracks.length-1];
+                            var stringAfterClinical = urlValue.substring(positionOfSubmit);
+                            var stringArrayAfterClinical = stringAfterClinical.split("&");
+                            var clinicalString = stringArrayAfterClinical[0] + "+" + extraTracks[extraTracks.length-1];
+                            var newUrl = urlValue.replace(stringArrayAfterClinical[0],clinicalString);
+//                            var newUrl = urlValue + "+" + extraTracks[extraTracks.length-1];
                             window.history.pushState({"html":window.location.html,"pageTitle":window.location.pageTitle},"", newUrl);
                         }
                         
@@ -1265,7 +1345,21 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                         $('#oncoprint-diagram-removeUCases-icon img')[0].attributes.src.value = 'images/removeUCases.svg';
                         $('#oncoprint-diagram-removeWhitespace-icon img')[0].attributes.src.value = 'images/removeWhitespace.svg';
                         utils.make_mouseover(d3.selectAll('.sample rect'),{linkage:true});        // hack =(
+                        
                         invokeDataManager();
+                        
+                        if(showPatients)
+                        {
+                            var AlteredPatientsNum= calculatePatientNum(PortalGlobals.getAlteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                            var UnalteredPatientsNum= calculatePatientNum(PortalGlobals.getUnalteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                            var totalPatientsNum = _.union(AlteredPatientsNum,UnalteredPatientsNum);
+                            var percentOfAlteredPatients = Math.ceil((AlteredPatientsNum.length/totalPatientsNum.length * 100).toFixed(1));
+                            $('#altered_value').text("Altered in "+ AlteredPatientsNum.length + " ("+ percentOfAlteredPatients +"%) of "+ totalPatientsNum.length +" patients/cases");
+                        }
+                        else
+                        {
+                            $('#altered_value').text("Altered in "+ PortalGlobals.getNumOfAlteredCases() + " ("+ Math.ceil(PortalGlobals.getPercentageOfAlteredCases()) +"%) of "+ PortalGlobals.getNumOfTotalCases() + " samples"); 
+                        }
                     }
                 });
             }
@@ -1339,7 +1433,7 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
         sortStatus[yMovement]=sortStatus[_dragElementIndex];
         sortStatus[_dragElementIndex] = tempSortStatus;
         
-        refreshOncoPrint(showPatientIdflat);
+        refreshOncoPrint(showPatients);
         zoom = reset_zoom();
         // sync
         oncoprint.zoom(zoom.val());
@@ -1802,8 +1896,22 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
             
             $('#oncoprint_diagram_topatientid_icon').click(function(){
               inner_loader_img.show();
+                if($('#switchPatientSample')[0].attributes.valuetype.value === "patients")
+                {
+                   $('#switchPatientSample').text("Show patients in OncoPrint"); 
+                   $('#switchPatientSample')[0].attributes.valuetype.value = "samples";
+                }
+                else
+                {
+                    $('#switchPatientSample').text("Show samples in OncoPrint")
+                    $('#switchPatientSample')[0].attributes.valuetype.value = "patients";
+                }
+                
               if($('#oncoprint_diagram_topatientid_icon img')[0].attributes.src.value === 'images/cool.svg')
-              {      
+              { 
+                var urlValueNow = window.location.href;
+                var newUrl = urlValueNow.replace("&show_samples=true","");
+                window.history.pushState({"html":window.location.html,"pageTitle":window.location.pageTitle},"", newUrl);
                 if($('#oncoprint_diagram_showmutationcolor_icon img')[0].attributes.src.value === 'images/mutationcolorsort.svg')
                 {
                     mutationColorControl = 'singleColor';
@@ -1826,10 +1934,10 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                 var totalPatientsNum = _.union(AlteredPatientsNum,UnalteredPatientsNum);
                 var percentOfAlteredPatients = Math.ceil((AlteredPatientsNum.length/totalPatientsNum.length * 100).toFixed(1));
                 
-                $('#altered_value').text("Altered in "+ AlteredPatientsNum.length + "("+ percentOfAlteredPatients +"%) of " + totalPatientsNum.length+ " patients/cases");
-                
-                showPatientIdflat = true;
-                refreshOncoPrint(showPatientIdflat);
+                $('#altered_value').text("Altered in "+ AlteredPatientsNum.length + " ("+ percentOfAlteredPatients +"%) of "+ totalPatientsNum.length +" cases/patients");
+
+                showPatients = true;
+                refreshOncoPrint(showPatients);
 
                 var zoomvalue = $('#oncoprint_zoom_slider')[0].value;
                 zoom = reset_zoom();
@@ -1853,10 +1961,14 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                 $('#oncoprint-diagram-showlegend-icon img')[0].attributes.src.value = 'images/showlegend.svg';
                 $('#oncoprint-diagram-removeUCases-icon img')[0].attributes.src.value = 'images/removeUCases.svg';
                 
-                $('.legend_missense_name').text("Mutation") ;
+                $('.legend_missense_name').text("Missense Mutation") ;
               }
                 else if($('#oncoprint_diagram_topatientid_icon img')[0].attributes.src.value === 'images/cool2.svg')
               {
+                var urlValueNow = window.location.href;
+                var newUrl = urlValueNow + "&show_samples=true";
+                window.history.pushState({"html":window.location.html,"pageTitle":window.location.pageTitle},"", newUrl);
+                
                 if($('#oncoprint_diagram_showmutationcolor_icon img')[0].attributes.src.value === 'images/mutationcolorsort.svg')
                 {
                     mutationColorControl = 'singleColor';
@@ -1872,10 +1984,10 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                     mutationColorControl = 'multiColor';
                     mutationColorSort = 'mutationcolorsort';  
                 }
-                showPatientIdflat = false;
+                showPatients = false;
                 refreshOncoPrint();
-                
-                $('#altered_value').text("Altered in "+ PortalGlobals.getNumOfAlteredCases() + "("+ Math.ceil(PortalGlobals.getPercentageOfAlteredCases()) +"%) of "+PortalGlobals.getNumOfTotalCases()+" samples");
+
+                $('#altered_value').text("Altered in "+ PortalGlobals.getNumOfAlteredCases() + " ("+ Math.ceil(PortalGlobals.getPercentageOfAlteredCases()) +"%) of "+ PortalGlobals.getNumOfTotalCases() + " samples");
                 
                 var zoomvalue = $('#oncoprint_zoom_slider')[0].value;
                 zoom = reset_zoom();
@@ -1903,7 +2015,134 @@ requirejs(  [         'Oncoprint',    'OncoprintUtils'],
                 $('.legend_nonmissense').css("display","inline");
               }
               inner_loader_img.hide();
+            });
+            
+            
+            
+            $('#switchPatientSample').click(function(){
+                inner_loader_img.show();
+                if($('#switchPatientSample')[0].attributes.valuetype.value === "patients")
+                {
+                   $('#switchPatientSample').text("Show patients in OncoPrint"); 
+                   $('#switchPatientSample')[0].attributes.valuetype.value = "samples";
+                }
+                else
+                {
+                    $('#switchPatientSample').text("Show samples in OncoPrint")
+                    $('#switchPatientSample')[0].attributes.valuetype.value = "patients";
+                }
+              
+              if($('#oncoprint_diagram_topatientid_icon img')[0].attributes.src.value === 'images/cool.svg')
+              {  
+                var urlValueNow = window.location.href;
+                var newUrl = urlValueNow.replace("&show_samples=true","");
+                window.history.pushState({"html":window.location.html,"pageTitle":window.location.pageTitle},"", newUrl);
+                if($('#oncoprint_diagram_showmutationcolor_icon img')[0].attributes.src.value === 'images/mutationcolorsort.svg')
+                {
+                    mutationColorControl = 'singleColor';
+                    mutationColorSort = 'mutationcolorsort';
+                }
+                else if($('#oncoprint_diagram_showmutationcolor_icon img')[0].attributes.src.value === 'images/colormutations.svg')
+                {
+                    mutationColorControl = 'multiColor';
+                    mutationColorSort = 'mutationcolornonsort';
+                }
+                else if($('#oncoprint_diagram_showmutationcolor_icon img')[0].attributes.src.value === 'images/uncolormutations.svg')
+                {                
+                    mutationColorControl = 'multiColor';
+                    mutationColorSort = 'mutationcolorsort';  
+                }
+                
+                var AlteredPatientsNum= calculatePatientNum(PortalGlobals.getAlteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                var UnalteredPatientsNum= calculatePatientNum(PortalGlobals.getUnalteredSampleIdList(),PortalGlobals.getPatientSampleIdMap());
+                
+                var totalPatientsNum = _.union(AlteredPatientsNum,UnalteredPatientsNum);
+                var percentOfAlteredPatients = Math.ceil((AlteredPatientsNum.length/totalPatientsNum.length * 100).toFixed(1));
+                
+                $('#altered_value').text("Altered in "+ AlteredPatientsNum.length + " ("+ percentOfAlteredPatients +"%) of "+ totalPatientsNum.length +" patients/cases");
+                
+                showPatients = true;
+                refreshOncoPrint(showPatients);
+
+                var zoomvalue = $('#oncoprint_zoom_slider')[0].value;
+                zoom = reset_zoom();
+                $('#oncoprint_zoom_slider')[0].value = zoomvalue;
+                // sync
+                oncoprint.zoom(zoomvalue);
+                
+                oncoprint.showUnalteredCases(!$('#toggle_unaltered_cases').is(":checked"));
+                if($('#oncoprint-diagram-removeWhitespace-icon img')[0].attributes.src.value === 'images/removeWhitespace.svg')
+                {
+                    oncoprint.toggleWhiteSpace(true);
+                }
+                else
+                {
+                    oncoprint.toggleWhiteSpace(false);
+                }
+                
+                utils.make_mouseover(d3.selectAll('.sample rect'),{linkage:true});        // hack =(
+                $('#oncoprint_diagram_topatientid_icon img')[0].attributes.src.value = 'images/cool2.svg';
+                
+                $('#oncoprint-diagram-showlegend-icon img')[0].attributes.src.value = 'images/showlegend.svg';
+                $('#oncoprint-diagram-removeUCases-icon img')[0].attributes.src.value = 'images/removeUCases.svg';
+                
+                $('.legend_missense_name').text("Missense Mutation") ;
+              }
+                else if($('#oncoprint_diagram_topatientid_icon img')[0].attributes.src.value === 'images/cool2.svg')
+              {
+                var urlValueNow = window.location.href;
+                var newUrl = urlValueNow + "&show_samples=true";
+                window.history.pushState({"html":window.location.html,"pageTitle":window.location.pageTitle},"", newUrl);
+                if($('#oncoprint_diagram_showmutationcolor_icon img')[0].attributes.src.value === 'images/mutationcolorsort.svg')
+                {
+                    mutationColorControl = 'singleColor';
+                    mutationColorSort = 'mutationcolorsort';
+                }
+                else if($('#oncoprint_diagram_showmutationcolor_icon img')[0].attributes.src.value === 'images/colormutations.svg')
+                {
+                    mutationColorControl = 'multiColor';
+                    mutationColorSort = 'mutationcolornonsort';
+                }
+                else if($('#oncoprint_diagram_showmutationcolor_icon img')[0].attributes.src.value === 'images/uncolormutations.svg')
+                {                
+                    mutationColorControl = 'multiColor';
+                    mutationColorSort = 'mutationcolorsort';  
+                }
+                showPatients = false;
+                refreshOncoPrint();
+
+                $('#altered_value').text("Altered in "+ PortalGlobals.getNumOfAlteredCases() + " ("+ Math.ceil(PortalGlobals.getPercentageOfAlteredCases()) +"%) of "+ PortalGlobals.getNumOfTotalCases() + " samples");
+
+                var zoomvalue = $('#oncoprint_zoom_slider')[0].value;
+                zoom = reset_zoom();
+                $('#oncoprint_zoom_slider')[0].value = zoomvalue;
+                // sync
+                oncoprint.zoom(zoomvalue);
+                
+                oncoprint.showUnalteredCases(!$('#toggle_unaltered_cases').is(":checked"));
+                if($('#oncoprint-diagram-removeWhitespace-icon img')[0].attributes.src.value === 'images/removeWhitespace.svg')
+                {
+                    oncoprint.toggleWhiteSpace(true);
+                }
+                else
+                {
+                    oncoprint.toggleWhiteSpace(false);
+                }
+                
+                utils.make_mouseover(d3.selectAll('.sample rect'),{linkage:true});        // hack =(
+                $('#oncoprint_diagram_topatientid_icon img')[0].attributes.src.value = 'images/cool.svg';
+                
+                $('#oncoprint-diagram-showlegend-icon img')[0].attributes.src.value = 'images/showlegend.svg';
+                $('#oncoprint-diagram-removeUCases-icon img')[0].attributes.src.value = 'images/removeUCases.svg';
+                
+                $('.legend_missense_name').text("Missense Mutation");
+                $('.legend_nonmissense').css("display","inline");
+              }
+              inner_loader_img.hide();
             });            
+            
+            
+            
             
             $('#oncoprint_diagram_topatientid_icon').hover(
             function () {